Added the as_wide_table function and new endpoints
The as_wide_table function combines all relational tables in austraits into one large table. It requires the purrr package to run (also added). I've added an endpoint for returning the entire wide table of austraits as a json object to test whether this is possible over the API. (It is not possible on my computer).