separate data prep from launch (#7)
This change separates the dataset preparation from the launch. 

We're storing data in an object store, as this is more reliable than
downloading from Zenodo. Previously, the code would download database,
turn it into a wide table, then create the means layer. However, this
creates a compute overhead at launch and also demands installing of the
AusTraits package.

By separating the data prep from the API launch, we prevent this
compute. Downside is that the dataset is larger to download.

@dcol2804 wrote the actual code for creating the means, I have only
split file into two. Hence listing as coauthored commit

Co-authored-by: David Coleman <50344360+dcol2804@users.noreply.github.com>